VC5Decompressor: struct RLV: use better-sized types.
We previously werent' fully using those size AND had padding.
With these sizes, we have no padding, and fully utilize them.

The perf appears to be slightly above the noise:
$ ~/rawspeed/build-Clang-Release/googlebenchmark/googlebenchmark-src/tools/compare.py -a benchmarks ~/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench{-prev,} --benchmark_display_aggregates_only=true --benchmark_counters_tabular=true --benchmark_repetitions=100 GOPR9172.GPR
RUNNING: /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench-prev --benchmark_display_aggregates_only=true --benchmark_counters_tabular=true --benchmark_repetitions=100 GOPR9172.GPR --benchmark_display_aggregates_only=true --benchmark_out=/tmp/tmpJnKeMv
2018-10-16 17:21:00
Running /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench-prev
Run on (8 X 4000 MHz CPU s)
CPU Caches:
  L1 Data 16K (x8)
  L1 Instruction 64K (x4)
  L2 Unified 2048K (x4)
  L3 Unified 8192K (x1)
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                        Time           CPU Iterations  CPUTime,s CPUTime/WallTime     Pixels Pixels/CPUTime Pixels/WallTime Raws/CPUTime Raws/WallTime WallTime,s
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
GOPR9172.GPR/threads:8/real_time_mean          320 ms        320 ms          2   0.319993          1.00002        12M       37.5016M        37.5025M      3.12513       3.12521   0.319985
GOPR9172.GPR/threads:8/real_time_median        319 ms        319 ms          2   0.319486          1.00003        12M       37.5603M        37.5611M      3.13003       3.13009    0.31948
GOPR9172.GPR/threads:8/real_time_stddev          1 ms          1 ms          2   1.44945m         13.6992u          0       168.656k         168.65k    0.0140547     0.0140542   1.44933m
RUNNING: /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench --benchmark_display_aggregates_only=true --benchmark_counters_tabular=true --benchmark_repetitions=100 GOPR9172.GPR --benchmark_display_aggregates_only=true --benchmark_out=/tmp/tmpOx9eVm
2018-10-16 17:22:04
Running /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench
Run on (8 X 4000 MHz CPU s)
CPU Caches:
  L1 Data 16K (x8)
  L1 Instruction 64K (x4)
  L2 Unified 2048K (x4)
  L3 Unified 8192K (x1)
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                        Time           CPU Iterations  CPUTime,s CPUTime/WallTime     Pixels Pixels/CPUTime Pixels/WallTime Raws/CPUTime Raws/WallTime WallTime,s
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
GOPR9172.GPR/threads:8/real_time_mean          318 ms        318 ms          2   0.318062          1.00003        12M       37.7286M        37.7296M      3.14405       3.14413   0.318054
GOPR9172.GPR/threads:8/real_time_median        318 ms        318 ms          2   0.317998          1.00003        12M       37.7361M        37.7371M      3.14468       3.14475    0.31799
GOPR9172.GPR/threads:8/real_time_stddev          0 ms          0 ms          2   467.753u         12.3211u          0       55.4371k        55.3923k     4.61976m      4.61603m   467.352u
Comparing /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench-prev to /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench
Benchmark                                                 Time             CPU      Time Old      Time New       CPU Old       CPU New
--------------------------------------------------------------------------------------------------------------------------------------
GOPR9172.GPR/threads:8/real_time_pvalue                 0.0000          0.0000      U Test, Repetitions: 100 vs 100
GOPR9172.GPR/threads:8/real_time_mean                  -0.0060         -0.0060           320           318           320           318
GOPR9172.GPR/threads:8/real_time_median                -0.0047         -0.0047           319           318           319           318
GOPR9172.GPR/threads:8/real_time_stddev                -0.6775         -0.6773             1             0             1             0

+++ src/librawspeed/decompressors/VC5Decompressor.cpp
@@ -43,8 +43,8 @@ namespace {
 struct RLV {
   const uint_fast8_t size; //!< Size of code word in bits
   const uint32_t bits;     //!< Code word bits right justified
-  const uint32_t count;    //!< Run length
-  const int32_t value;     //!< Run value (unsigned)
+  const uint16_t count;    //!< Run length
+  const uint8_t value;     //!< Run value (unsigned)
 };
 #define RLVTABLE(n)                                                            \
   struct {                                                                     \
